requirements.txt
garminconnect
pandas
python-dotenv